test(inductor): 更新测试用例以使用float16数据类型
修改测试用例中的模型初始化，添加dtype=torch.float16参数以测试模型在float16数据类型下的行为

+++ examples/inductor_pattern_match/add_layernorm.py
@@ -133,7 +133,7 @@ class AddLayerNorm(torch.nn.Module):
 def test_add_layernorm_pattern():
     """测试Add + LayerNorm模式匹配"""
 
-    model = AddLayerNorm().to(device)
+    model = AddLayerNorm().to(device, dtype=torch.float16)
     compiled_model = torch.compile(model, backend="inductor")
 
     # 创建测试数据
